pyproject.toml
[project]
name = "clubs"
version = "0.1.0"
description = ""
authors = [{ name = "one-zero-eight" }]
urls = { homepage = "https://t.me/one_zero_eight" }
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "authlib>=1.6.5",
    "colorlog>=6.8.2",
    "cryptography>=44.0.0",
    "fastapi[standard]>=0.115.6",
    "fastapi-swagger>=0.2.3",
    "fastapi-derive-responses>=0.1.4",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.2",
    "ruff>=0.9.2",
    "uvicorn>=0.34.0",
    "beanie>=1.30.0,<2.0.0",
    "pyvips>=3.0.0",
    "python-magic>=0.4.27",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "astor>=0.8.1",
    "pygments>=2.18.0",
    "simple-term-menu>=1.6.6",
]

[tool.ruff]
line-length = 120
lint.extend-select = ["I", "UP", "PL"]
lint.extend-ignore = ["PLC0415", "PLR"]
target-version = "py313"
